requirements.txt
pandas>=2.0.0
matplotlib>=3.7.0
langchain>=0.1.0
openai>=1.0.0
langgraph>=0.0.17
streamlit>=1.30.0
pydantic>=2.0.0
python-dotenv>=1.0.0
sqlalchemy>=2.0.0
scikit-learn>=1.2.0
langchain_community
seaborn
json5